requirements.txt

PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
glfw==2.7.0
numpy==1.26.4
pyrr==0.10.3